requirements.txt
faiss-cpu
sentence-transformers
langchain
pymupdf
python-docx
pymysql
numpy
